Move writeBlock function outside of class.

+++ libraries/EEPROM/examples/Example2_AllFunctions/Example2_AllFunctions.ino
@@ -30,7 +30,7 @@ void setup()
 
   long startTime;
   long endTime;
-  uint16_t randomLocation;
+  int randomLocation;
 
   //Test erase time
   startTime = millis();
